refactor: split a pressHandle function into subfunctions
