+++ src/index.js
@@ -1,6 +1,6 @@
 import './styles/index.scss';
 import init from './js/intitKeys';
-import pressHandle from './js/pressHandle';
+import pressHandle from './js/pressFuncs/pressHandle';
 import keyboardStore from './store/store';
 
 window.addEventListener('DOMContentLoaded', init);
