+++ src/js/pressFuncs/btnTypeSwitches.js
